+++ common/lib-sample-dagger/src/main/java/ru/surfstudio/android/sample/dagger/ui/base/dagger/screen/DefaultActivityScreenModule.kt
@@ -1,26 +1,21 @@
 package ru.surfstudio.android.sample.dagger.ui.base.dagger.screen
 
-import android.content.SharedPreferences
 import dagger.Module
 import dagger.Provides
 import ru.surfstudio.android.core.mvp.scope.ActivityViewPersistentScope
 import ru.surfstudio.android.core.ui.event.ScreenEventDelegateManager
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigatorForActivity
-import ru.surfstudio.android.core.ui.permission.PermissionManager
-import ru.surfstudio.android.core.ui.permission.PermissionManagerForActivity
 import ru.surfstudio.android.core.ui.provider.ActivityProvider
 import ru.surfstudio.android.core.ui.scope.ActivityPersistentScope
 import ru.surfstudio.android.core.ui.scope.ScreenPersistentScope
 import ru.surfstudio.android.core.ui.state.ScreenState
 import ru.surfstudio.android.dagger.scope.PerScreen
-import ru.surfstudio.android.sample.dagger.ui.base.error.DefaultErrorHandlerModule
 import ru.surfstudio.android.message.DefaultMessageController
 import ru.surfstudio.android.message.MessageController
 import ru.surfstudio.android.mvp.dialog.navigation.navigator.DialogNavigator
 import ru.surfstudio.android.mvp.dialog.navigation.navigator.DialogNavigatorForActivity
-import ru.surfstudio.android.shared.pref.NO_BACKUP_SHARED_PREF
-import javax.inject.Named
+import ru.surfstudio.android.sample.dagger.ui.base.error.DefaultErrorHandlerModule
 
 @Module(includes = [DefaultErrorHandlerModule::class])
 class DefaultActivityScreenModule(private val activityViewPersistentScope: ActivityViewPersistentScope) : DefaultScreenModule() {
@@ -50,22 +45,6 @@ class DefaultActivityScreenModule(private val activityViewPersistentScope: Activ
             eventDelegateManager: ScreenEventDelegateManager
     ): ActivityNavigator {
         return ActivityNavigatorForActivity(activityProvider, eventDelegateManager)
-    }
-
-    @Provides
-    @PerScreen
-    internal fun providePermissionManager(
-            eventDelegateManager: ScreenEventDelegateManager,
-            activityNavigator: ActivityNavigator,
-            @Named(NO_BACKUP_SHARED_PREF) sharedPreferences: SharedPreferences,
-            activityProvider: ActivityProvider
-    ): PermissionManager {
-        return PermissionManagerForActivity(
-                eventDelegateManager,
-                activityNavigator,
-                sharedPreferences,
-                activityProvider
-        )
     }
 
     @Provides