+++ common/lib-sample-dagger/src/main/java/ru/surfstudio/android/sample/dagger/ui/base/dagger/widget/DefaultWidgetScreenModule.kt
@@ -1,6 +1,5 @@
 package ru.surfstudio.android.sample.dagger.ui.base.dagger.widget
 
-import android.content.SharedPreferences
 import dagger.Module
 import dagger.Provides
 import ru.surfstudio.android.core.ui.ScreenType
@@ -8,9 +7,6 @@ import ru.surfstudio.android.core.ui.event.ScreenEventDelegateManager
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigatorForActivity
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigatorForFragment
-import ru.surfstudio.android.core.ui.permission.PermissionManager
-import ru.surfstudio.android.core.ui.permission.PermissionManagerForActivity
-import ru.surfstudio.android.core.ui.permission.PermissionManagerForFragment
 import ru.surfstudio.android.core.ui.provider.ActivityProvider
 import ru.surfstudio.android.core.ui.provider.FragmentProvider
 import ru.surfstudio.android.core.ui.scope.ScreenPersistentScope
@@ -26,7 +22,6 @@ import ru.surfstudio.android.mvp.widget.scope.WidgetViewPersistentScope
 import ru.surfstudio.android.mvp.widget.state.WidgetScreenState
 import ru.surfstudio.android.sample.dagger.ui.base.dagger.screen.DefaultScreenModule
 import ru.surfstudio.android.sample.dagger.ui.base.error.DefaultErrorHandlerModule
-import ru.surfstudio.android.shared.pref.NO_BACKUP_SHARED_PREF
 import javax.inject.Named
 
 private const val PARENT_TYPE_DAGGER_NAME = "parent_type"
@@ -93,32 +88,6 @@ class DefaultWidgetScreenModule(
     @PerScreen
     internal fun provideEventDelegateManager(): ScreenEventDelegateManager {
         return persistentScope.screenEventDelegateManager
-    }
-
-    @Provides
-    @PerScreen
-    internal fun providePermissionManager(
-            eventDelegateManager: ScreenEventDelegateManager,
-            activityProvider: ActivityProvider,
-            activityNavigator: ActivityNavigator,
-            @Named(PARENT_TYPE_DAGGER_NAME) parentType: ScreenType,
-            @Named(NO_BACKUP_SHARED_PREF) sharedPreferences: SharedPreferences
-    ): PermissionManager {
-        return if (parentType == ScreenType.FRAGMENT)
-            PermissionManagerForFragment(
-                    eventDelegateManager,
-                    activityProvider,
-                    activityNavigator,
-                    sharedPreferences,
-                    createFragmentProvider()
-            )
-        else
-            PermissionManagerForActivity(
-                    eventDelegateManager,
-                    activityNavigator,
-                    sharedPreferences,
-                    activityProvider
-            )
     }
 
     @Provides