+++ navigation/lib-navigation-observer/src/main/java/ru/surfstudio/android/navigation/observer/route/ResultRoute.kt
@@ -3,4 +3,7 @@ package ru.surfstudio.android.navigation.observer.route
 import ru.surfstudio.android.navigation.route.Route
 import java.io.Serializable
 
-interface ResultRoute<T : Serializable> : Route
+interface ResultRoute<T : Serializable> : Route {
+
+    fun getId(): String
+}