+++ permission/lib-permission/src/main/java/ru/surfstudio/android/core/ui/permission/screens/default_permission_rational/DefaultPermissionRationalActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import ru.surfstudio.android.core.ui.permission.R
-import ru.surfstudio.android.core.ui.navigation.Route
+import ru.surfstudio.android.navigation.route.activity.getDataBundle
 
 /**
  * Экран объяснения причины запроса разрешения.
@@ -12,17 +12,19 @@ import ru.surfstudio.android.core.ui.navigation.Route
 class DefaultPermissionRationalActivity : AppCompatActivity() {
 
     private val permissionsRationalStr: String
-        get() = intent.getStringExtra(Route.EXTRA_FIRST) ?: ""
+        get() = DefaultPermissionRationalRoute(intent.getDataBundle()!!).permissionRationalStr
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         AlertDialog
-                .Builder(this)
-                .setMessage(permissionsRationalStr)
-                .setNeutralButton(R.string.default_permission_rational_got_it, null)
-                .setOnDismissListener { finish() }
-                .create()
-                .show()
+            .Builder(this)
+            .setMessage(permissionsRationalStr)
+            .setNeutralButton(R.string.default_permission_rational_got_it, null)
+            .setOnDismissListener {
+                finish()
+            }
+            .create()
+            .show()
     }
 }