+++ permission/lib-permission/src/main/java/ru/surfstudio/android/core/ui/permission/screens/default_permission_rational/DefaultPermissionRationalRoute.kt
@@ -1,19 +1,33 @@
 package ru.surfstudio.android.core.ui.permission.screens.default_permission_rational
 
-import android.content.Context
+import android.app.Activity
 import android.content.Intent
-import ru.surfstudio.android.core.ui.navigation.Route
-import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityWithResultRoute
+import android.os.Bundle
+import androidx.core.os.bundleOf
+import ru.surfstudio.android.navigation.observer.route.ActivityWithResultRoute
+import ru.surfstudio.android.navigation.route.Route
 import java.io.Serializable
 
 /**
  * Маршрут экрана объяснения причины запроса разрешения.
  */
 class DefaultPermissionRationalRoute(
-        private val permissionRationalStr: String
+    val permissionRationalStr: String
 ) : ActivityWithResultRoute<Serializable>() {
 
-    override fun prepareIntent(context: Context?): Intent =
-            Intent(context, DefaultPermissionRationalActivity::class.java)
-                    .apply { putExtra(Route.EXTRA_FIRST, permissionRationalStr) }
+    constructor(args: Bundle) : this(
+        args[Route.EXTRA_FIRST] as String
+    )
+
+    override fun getScreenClass() = DefaultPermissionRationalActivity::class.java
+
+    override fun prepareData() = bundleOf(
+        Route.EXTRA_FIRST to permissionRationalStr
+    )
+
+    override val uniqueId: String = permissionRationalStr
+
+    override fun parseResultIntent(resultCode: Int, resultIntent: Intent?): Boolean {
+        return resultCode == Activity.RESULT_OK
+    }
 }