+++ permission/sample/src/main/java/ru/surfstudio/android/permission/sample/screen/main/camera/CameraPermissionRequest.kt
@@ -6,4 +6,8 @@ import ru.surfstudio.android.core.ui.permission.PermissionRequest
 class CameraPermissionRequest : PermissionRequest() {
     override val permissions: Array<String>
         get() = arrayOf(Manifest.permission.CAMERA)
+
+    init {
+        showPermissionsRational = true
+    }
 }