+++ permission/sample/src/main/java/ru/surfstudio/android/permission/sample/screen/main/camera/CameraRoute.kt
@@ -7,7 +7,12 @@ import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityRoute
 
 class CameraRoute : ActivityRoute() {
 
-    override fun prepareIntent(context: Context?): Intent {
-        return Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+    override fun prepareIntent(context: Context): Intent {
+        val takePictureIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+
+        return Intent.createChooser(
+            takePictureIntent,
+            "Choose image"
+        )
     }
 }