+++ permission/RELEASE_NOTES.md
@@ -9,3 +9,6 @@
 from `ru.surfstudio.android.core.ui.event.result` to:  `ru.surfstudio.android.core.ui.navigation.event.result`
 * ANDDEP-1048 Fixing wrong docs links and docs structure
 * ANDDEP-1148 code updated for target sdk 30
+* **NO BACKWARD COMPATIBILITY** ANDDEP-1201 removed core-navigation from dependencies, instead user
+  new navigation module. Now there is only one PermissionManager which should be used as singleton.
+  Added check for grant once.